refactor(mcp): restore Resolve as the case entry point and split out tools.go
